+++ src/utils/toolbar.ts
@@ -7,28 +7,23 @@ export const getCurrentStructure = (
     | ShapeEnum
     | {
         title: string
-        img: string
         tools: ShapeEnum[]
         vertical: boolean
       }
   )[]
 ) => {
   return _.flow(
-    _.map(
-      (
-        group: ShapeEnum | { title: string; vertical: boolean; img: string; tools: ShapeEnum[] }
-      ) => {
-        if (!_.isObject(group)) {
-          if (_.includes(group, availableTools)) {
-            return group
-          }
-          return null
+    _.map((group: ShapeEnum | { title: string; vertical: boolean; tools: ShapeEnum[] }) => {
+      if (!_.isObject(group)) {
+        if (_.includes(group, availableTools)) {
+          return group
         }
-        const tools = _.intersection(availableTools, group.tools)
-        if (!tools) return null
-        return { ...group, tools }
+        return null
       }
-    ),
+      const tools = _.intersection(availableTools, group.tools)
+      if (!tools) return null
+      return { ...group, tools }
+    }),
     _.compact
   )(defaultStructure)
 }